move checkA11y function to support/helpers
