+++ cypress/integration/pages/testsForAllPages.js
@@ -1,7 +1,7 @@
 /* eslint-disable cypress/no-unnecessary-waiting */
 // For testing important features that differ between services, e.g. Timestamps.
 // We recommend using inline conditional logic to limit tests to services which differ.
-import checkA11y from './checkA11y';
+import checkA11y from '../../support/helpers/checkA11y';
 import getDataUrl from '../../support/helpers/getDataUrl';
 
 export const testsThatAlwaysRunForAllPages = ({ service, pageType }) => {
